Added new controllers and moved other controllers.
