+++ FribergFastigheter.Server/Controllers/BrokerFirmApi/BrokerController.cs
@@ -8,9 +8,9 @@ using System.ComponentModel.DataAnnotations;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
-namespace FribergFastigheter.Server.Controllers.Broker
+namespace FribergFastigheter.Server.Controllers.BrokerFirmApi
 {
-    [Route("api/Brokerfirm/Housing")]
+    [Route("api/BrokerFirm/Broker")]
     [ApiController]
     public class BrokerController : ControllerBase
     {
