+++ FribergFastigheter.Server/Controllers/HousingApi/HousingBrokerController.cs
@@ -1,0 +1,43 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+
+namespace FribergFastigheter.Server.Controllers.HousingApi
+{
+	[Route("api/Housing/Broker")]
+	[ApiController]
+	public class HousingBrokerController : ControllerBase
+	{
+		// GET: api/<HousingBrokerController>
+		[HttpGet]
+		public IEnumerable<string> Get()
+		{
+			return new string[] { "value1", "value2" };
+		}
+
+		// GET api/<HousingBrokerController>/5
+		[HttpGet("{id}")]
+		public string Get(int id)
+		{
+			return "value";
+		}
+
+		// POST api/<HousingBrokerController>
+		[HttpPost]
+		public void Post([FromBody] string value)
+		{
+		}
+
+		// PUT api/<HousingBrokerController>/5
+		[HttpPut("{id}")]
+		public void Put(int id, [FromBody] string value)
+		{
+		}
+
+		// DELETE api/<HousingBrokerController>/5
+		[HttpDelete("{id}")]
+		public void Delete(int id)
+		{
+		}
+	}
+}
