+++ FribergFastigheter.Server/Controllers/HousingApi/HousingBrokerFirmController.cs
@@ -1,0 +1,43 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+// For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
+
+namespace FribergFastigheter.Server.Controllers.HousingApi
+{
+	[Route("api/Housing/BrokerFirm")]
+	[ApiController]
+	public class HousingBrokerFirmController : ControllerBase
+	{
+		// GET: api/<HousingBrokerFirmController>
+		[HttpGet]
+		public IEnumerable<string> Get()
+		{
+			return new string[] { "value1", "value2" };
+		}
+
+		// GET api/<HousingBrokerFirmController>/5
+		[HttpGet("{id}")]
+		public string Get(int id)
+		{
+			return "value";
+		}
+
+		// POST api/<HousingBrokerFirmController>
+		[HttpPost]
+		public void Post([FromBody] string value)
+		{
+		}
+
+		// PUT api/<HousingBrokerFirmController>/5
+		[HttpPut("{id}")]
+		public void Put(int id, [FromBody] string value)
+		{
+		}
+
+		// DELETE api/<HousingBrokerFirmController>/5
+		[HttpDelete("{id}")]
+		public void Delete(int id)
+		{
+		}
+	}
+}
